// api/model/updateUserCommand.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateUserCommand { 
    userId?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    streetAddress?: string | null;
    postalCode?: string | null;
    city?: string | null;
    state?: string | null;
}